trambu-infrastructure/src/main/java/be/doji/productivity/trambu/infrastructure/repository/ActivityRepository.java
/**
 * GNU Affero General Public License Version 3
 *
 * Copyright (c) 2019 Stijn Dejongh
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU Affero General Public License for more details.
 *
 * You should have received a copy of the GNU Affero General Public License
 * along with this program. If not, see <http://www.gnu.org/licenses/>.
 */
package be.doji.productivity.trambu.infrastructure.repository;

import be.doji.productivity.trambu.domain.activity.Activity;
import java.util.ArrayList;
import java.util.List;
import org.springframework.stereotype.Service;

@Service
public class ActivityRepository {

  private List<Activity> activities = new ArrayList<>();

  public List<Activity> getAll() {
    return new ArrayList<>(this.activities);
  }

  public void save(Activity activityToSave) {
    this.activities.add(activityToSave);
  }

  public void clear() {
    this.activities.clear();
  }
}
